feat: add multi-input mapping for physics and optimize CI linting
- Implemented `ResolveInput` in `BasePhysicsScheme` to allow schemes to consume
  fields from both import and export states (enabling scheme chaining).
- Added `input_mapping` and `output_mapping` support for physics schemes via YAML configuration.
- Updated CI linter (`clang-tidy`) to analyze only project source files and exclude dependencies.
- Fixed several logical expression corruptions (e.g., `!= nullptr != nullptr`).
- Improved code quality by replacing `std::endl` with `\n`, using `nullptr`, and adding explicit braces.
- Updated `AcesConfig` and parser to support unified field name resolution.

+++ src/physics/aces_dust.cpp
@@ -49,9 +49,9 @@ void DustScheme::Run(AcesImportState& import_state, AcesExportState& export_stat
         dust_emis.data() == nullptr)
         return;
 
-    int nx = dust_emis.extent(0);
-    int ny = dust_emis.extent(1);
-    int nz = dust_emis.extent(2);
+    int nx = static_cast<int>(dust_emis.extent(0));
+    int ny = static_cast<int>(dust_emis.extent(1));
+    int nz = static_cast<int>(dust_emis.extent(2));
 
     const double CH_DUST = 9.375e-10;  // Default tuning factor
     double u_ts0_const = u_ts0_;
